Revert "SKETCH-2532: Fixes for setting the color scheme via SketcherWidget::setColorScheme (#65)"
This reverts commit bc07b0a1093625f67f7f03ba021abe131a6699ba.
