+++ src/schrodinger/sketcher/dialog/rendering_settings_dialog.cpp
@@ -93,9 +93,6 @@ void RenderingSettingsDialog::loadSettingsFromModel()
     // When loading from the model, we don't want to the updates to be reflected
     // back to the model
     loadSettings(settings);
-    // modifying the widget values started the update timer, but we don't need
-    // to update the model since that's where these values just came from
-    m_update_timer->stop();
 }
 
 void RenderingSettingsDialog::updateWidgets()
