+++ src/schrodinger/sketcher/molviewer/bond_display_settings.cpp
@@ -3,7 +3,6 @@
 #include <stdexcept>
 #include <stdio.h>
 
-#include "schrodinger/sketcher/model/sketcher_model.h"
 #include "schrodinger/sketcher/molviewer/constants.h"
 
 namespace schrodinger
@@ -14,11 +13,6 @@ namespace sketcher
 qreal static scale_value(qreal scale, qreal default_value, qreal scaling_factor)
 {
     return (1 + scaling_factor * (scale - 1)) * default_value;
-}
-
-BondDisplaySettings::BondDisplaySettings()
-{
-    setColorScheme(ColorScheme::DEFAULT);
 }
 
 void BondDisplaySettings::setScale(qreal scale)
@@ -33,16 +27,12 @@ void BondDisplaySettings::setScale(qreal scale)
                                  BOND_HASH_SPACING_SCALING_FACTOR);
 }
 
-void BondDisplaySettings::setColorScheme(const ColorScheme& scheme,
-                                         const QColor& carbon_color)
+void BondDisplaySettings::setColorScheme(const ColorScheme& scheme)
 {
-    // update the annotation color
-    AbstractAtomOrBondDisplaySettings::setColorScheme(scheme, carbon_color);
-    // the bond color should match the carbon color
-    RDKit::ColourPalette temp_color_palette;
-    populatePaletteForColorScheme(scheme, carbon_color, temp_color_palette);
-    m_color = getAtomColorFromPalette(static_cast<int>(Element::C),
-                                      temp_color_palette);
+    m_annotation_color =
+        (scheme == ColorScheme::DARK_MODE || scheme == ColorScheme::WHITE_BLACK)
+            ? ANNOTATION_COLOR_DARK
+            : ANNOTATION_COLOR;
 }
 
 } // namespace sketcher
