+++ src/schrodinger/sketcher/molviewer/bond_display_settings.h
@@ -3,10 +3,8 @@
 #include <QtGlobal>
 #include <QColor>
 
-#include "schrodinger/sketcher/definitions.h"
-#include "schrodinger/sketcher/image_constants.h"
-#include "schrodinger/sketcher/molviewer/abstract_atom_or_bond_display_settings.h"
 #include "schrodinger/sketcher/molviewer/constants.h"
+#include "schrodinger/sketcher/image_constants.h"
 
 namespace schrodinger
 {
@@ -16,11 +14,10 @@ namespace sketcher
 /**
  * An object used to store all settings relevant to BondItems
  */
-class SKETCHER_API BondDisplaySettings
-    : public AbstractAtomOrBondDisplaySettings
+class BondDisplaySettings
 {
   public:
-    BondDisplaySettings();
+    BondDisplaySettings() = default;
     // Prevent implicit copies, since that could cause bugs, as the Scene shares
     // a single BondDisplaySettings object with all BondItems.
     explicit BondDisplaySettings(const BondDisplaySettings&) = default;
@@ -33,9 +30,11 @@ class SKETCHER_API BondDisplaySettings
     /// The distance between hash marks in "down" (i.e. dashed) wedges
     qreal m_hash_spacing = DEFAULT_BOND_HASH_SPACING;
     /// The bond color
-    QColor m_color;
+    QColor m_color = Qt::black;
     /// Whether to display stereochemistry labels when present
     bool m_stereo_labels_shown = true;
+    /// The color of the annotation text
+    QColor m_annotation_color = ANNOTATION_COLOR;
 
     /**
      * Scale bond width by the specified value.  Note that double bond spacing
@@ -52,8 +51,7 @@ class SKETCHER_API BondDisplaySettings
      * color of annotations (which is lighter on dark modes)
      * @param scheme The color scheme to use
      */
-    void setColorScheme(const ColorScheme& scheme,
-                        const QColor& carbon_color = QColor()) override;
+    void setColorScheme(const ColorScheme& scheme);
 };
 } // namespace sketcher
 } // namespace schrodinger
